WIP - markup refactored and element placed around the header which will have the blue background classes turned on if it's a topic finder.
Layout change triggered by adding the query '&topic-finder=a' to the URL.

Bespoke CSS added to tweak the layout - specifically to remove the thin
blue bar and the spacing between the breadcrumb and heading elements.

I think the layout is good - it's now just getting the list of topics
(eg Education) to allow them to replace the current filler text of
'thing'.